feat : 추가 개발

+++ src/main/java/gangdong/diet/domain/review/service/ReviewServiceImpl.java
@@ -1,10 +1,13 @@
 package gangdong.diet.domain.review.service;
 
+import gangdong.diet.domain.member.entity.Member;
+import gangdong.diet.domain.member.repository.MemberRepository;
 import gangdong.diet.domain.post.entity.Post;
 import gangdong.diet.domain.post.repository.PostRepository;
 import gangdong.diet.domain.review.dto.ReviewRequest;
 import gangdong.diet.domain.review.entity.Review;
 import gangdong.diet.domain.review.repository.ReviewRepository;
+import gangdong.diet.global.auth.MemberDetails;
 import gangdong.diet.global.exception.ApiException;
 import gangdong.diet.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -18,16 +21,21 @@ public class ReviewServiceImpl implements ReviewService{
 
     private final ReviewRepository reviewRepository;
     private final PostRepository postRepository;
+    private final MemberRepository memberRepository;
 
     @Transactional
     @Override
-    public void addReview(ReviewRequest reviewRequest) {
-        Post post = postRepository.findById(reviewRequest.getPostId())
+    public void addReview(Long postId, ReviewRequest reviewRequest, MemberDetails memberDetails) {
+        Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ApiException(ErrorCode.POST_NOT_FOUND));
+
+        Member member = memberRepository.findByMemberEmail(memberDetails.getUsername())
+                .orElseThrow(() -> new ApiException(ErrorCode.MEMBER_NOT_FOUND));
 
         Review review = Review.builder()
                 .content(reviewRequest.getContent())
                 .rating(reviewRequest.getRating())
+                .member(member)
                 .post(post)
                 .build();
 
@@ -38,20 +46,37 @@ public class ReviewServiceImpl implements ReviewService{
 
     @Transactional
     @Override
-    public void updateReview(Long reviewId, ReviewRequest reviewRequest) {
+    public void updateReview(Long reviewId, ReviewRequest reviewRequest, MemberDetails memberDetails) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new ApiException(ErrorCode.REVIEW_NOT_FOUND));
 
-        review.setContent(reviewRequest.getContent());
+        Member member = memberRepository.findByMemberEmail(memberDetails.getUsername())
+                .orElseThrow(() -> new ApiException(ErrorCode.MEMBER_NOT_FOUND));
+
+        if (member.getId() == review.getMember().getId()) {
+            review.setContent(reviewRequest.getContent());
+        }
+        else {
+            throw new ApiException(ErrorCode.BAD_REQUEST);
+        }
+
 
     }
 
     @Transactional
     @Override
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(Long reviewId, MemberDetails memberDetails) {
         Review review = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new ApiException(ErrorCode.REVIEW_NOT_FOUND));
 
-        reviewRepository.delete(review);
+        Member member = memberRepository.findByMemberEmail(memberDetails.getUsername())
+                .orElseThrow(() -> new ApiException(ErrorCode.MEMBER_NOT_FOUND));
+
+        if (member.getId() == review.getMember().getId()) {
+            reviewRepository.delete(review);
+        }
+        else {
+            throw new ApiException(ErrorCode.BAD_REQUEST);
+        }
     }
 }
